refactor: organizado o módulo de perfil
