+++ lib/src/perfil/components/new_tag_dialog.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:gastos/Model/Tag.dart';
+import 'package:gastos/ModelHelper/DatabaseHelper.dart';
+
+void adicionarTag(String newTagName) {
+  DatabaseHelper databaseHelper = DatabaseHelper();
+  Tag newTag = Tag(name: newTagName);
+  databaseHelper.insertTag(newTag);
+}
+
+displayNewTagDialog(BuildContext context, _tagController) async {
+  String? newTag = await showDialog<String>(
+    context: context,
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text('Adicionar Tag'),
+        content: TextField(
+          controller: _tagController,
+          onChanged: (value) {
+            // Você pode adicionar validações ou manipular o valor conforme necessário
+          },
+          decoration:
+          const InputDecoration(hintText: 'Digite o nome da nova tag'),
+        ),
+        actions: <Widget>[
+          TextButton(
+            child: const Text('Adicionar'),
+            onPressed: () {
+              if (_tagController.text.isNotEmpty) {
+                adicionarTag(_tagController.text);
+              }
+              Navigator.of(context).pop();
+            },
+          ),
+        ],
+      );
+    },
+  );
+
+  print('sai do dialogo');
+  print('newTag: $_tagController');
+  _tagController.clear();
+}
