+++ lib/src/perfil/perfil_module.dart
@@ -1,0 +1,122 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+int globalIndex = 0;
+
+void _nomeChange(String value) {
+  if (value.isEmpty) {
+    value = '';
+  } else {
+    value = value;
+  }
+  print('nome: $value');
+  saveNome(value);
+}
+
+void _salarioChange(String value) {
+  if (value.isEmpty) {
+    value = '0';
+  } else {
+    value = value.replaceAll(RegExp(r'[R\$]'), '');
+  }
+  print('salario: $value');
+  saveSalario(value);
+}
+
+Future<void> saveSalario(String salario) async {
+  final prefs = await SharedPreferences.getInstance();
+
+  if (RegExp(r'^\d+(\.\d+)?$').hasMatch(salario)) {
+    prefs.setDouble('salario', double.parse(salario));
+    double? salarioSalvo = prefs.getDouble('salario');
+    print('salario salvo: $salarioSalvo');
+  } else {
+    print('Formato de salário inválido');
+  }
+}
+
+Future<void> saveNome(String nome) async {
+  final prefs = await SharedPreferences.getInstance();
+  prefs.setString('nome', nome);
+  String? nomeSalvo = prefs.getString('nome');
+  print('nome salvo: $nomeSalvo');
+}
+
+Future<String> getSalario() async {
+  final prefs = await SharedPreferences.getInstance();
+  double? salarioSalvo = prefs.getDouble('salario');
+  print('salario salvo recuperado: $salarioSalvo');
+  if (salarioSalvo == null) {
+    return 'R\$0.0';
+  }
+  String salarioSalvoString = salarioSalvo.toString();
+  return 'R\$$salarioSalvoString';
+}
+
+Future<String> getNome() async {
+  final prefs = await SharedPreferences.getInstance();
+  String? nomeSalvo = prefs.getString('nome');
+  print('nome salvo recuperado: $nomeSalvo');
+  if (nomeSalvo == null) {
+    return 'Nome do Usuário';
+  }
+  return nomeSalvo;
+}
+
+Widget getSalarioTextField() {
+  return FutureBuilder(
+    future: getSalario(),
+    builder: (context, AsyncSnapshot<String> snapshot) {
+      if (snapshot.hasData) {
+        return TextFormField(
+          initialValue: snapshot.data,
+          onChanged: (value) => _salarioChange(value),
+          keyboardType: TextInputType.number,
+          decoration: InputDecoration(border: OutlineInputBorder()),
+        );
+      } else {
+        return const Text(
+          'Saldo: \$0.0',
+          style: TextStyle(color: Colors.black),
+        );
+      }
+    },
+  );
+}
+
+Widget getNomeTextField() {
+  return FutureBuilder(
+    future: getNome(),
+    builder: (context, AsyncSnapshot<String> snapshot) {
+      if (snapshot.hasData) {
+        return TextFormField(
+          initialValue: snapshot.data,
+          onChanged: (value) => _nomeChange(value),
+          keyboardType: TextInputType.name,
+          decoration: InputDecoration(border: OutlineInputBorder()),
+        );
+      } else {
+        return const Text(
+          'Saldo: \$0.0',
+          style: TextStyle(color: Colors.black),
+        );
+      }
+    },
+  );
+}
+
+Widget getDefaultTagsWidgets() {
+  Map<String, Widget> tags = {
+    'gasolina' : Icon(Icons.local_gas_station),
+    'comida' : Icon(Icons.restaurant),
+    'gasto' : Icon(Icons.paid)
+  };
+  List<Widget> tagWidgets = [];
+  for (var tag in tags.values) {
+    final padding =
+    Padding(padding: EdgeInsets.symmetric(horizontal: 5), child: tag);
+    tagWidgets.add(padding);
+  }
+  return Row(children: tagWidgets);
+}
